Rename Sink mixin to Pipe, since that's what it does
