+++ lib/combinator/accumulate.js
@@ -3,7 +3,7 @@
 /** @author John Hann */
 
 var Stream = require('../Stream');
-var Sink = require('../sink/Sink');
+var Sink = require('../sink/Pipe');
 var cons = require('./build').cons;
 var drain = require('./observe').drain;
 var noop = require('../base').noop;
