+++ lib/combinator/combine.js
@@ -3,7 +3,7 @@
 /** @author John Hann */
 
 var Stream = require('../Stream');
-var Sink = require('../sink/Sink');
+var Sink = require('../sink/Pipe');
 var IndexSink = require('../sink/IndexSink');
 var CompoundDisposable = require('../disposable/CompoundDisposable');
 var base = require('../base');
