+++ lib/combinator/timestamp.js
@@ -3,7 +3,7 @@
 /** @author John Hann */
 
 var Stream = require('../Stream');
-var Sink = require('../sink/Sink');
+var Sink = require('../sink/Pipe');
 
 exports.timestamp = timestamp;
