+++ lib/combinator/transform.js
@@ -4,7 +4,7 @@
 
 var Stream = require('../Stream');
 var flatMap = require('./join').flatMap;
-var Sink = require('../sink/Sink');
+var Sink = require('../sink/Pipe');
 
 exports.map = map;
 exports.ap  = ap;
