+++ lib/sink/IndexSink.js
@@ -2,7 +2,7 @@
 /** @author Brian Cavalier */
 /** @author John Hann */
 
-var Sink = require('../sink/Sink');
+var Sink = require('./Pipe');
 
 module.exports = IndexSink;
