+++ lib/sink/Pipe.js
@@ -1,0 +1,27 @@
+/** @license MIT License (c) copyright 2010-2014 original author or authors */
+/** @author Brian Cavalier */
+/** @author John Hann */
+
+module.exports = Pipe;
+
+/**
+ * A sink mixin that simply forwards event, end, and error to
+ * another sink.
+ * @param sink
+ * @constructor
+ */
+function Pipe(sink) {
+	this.sink = sink;
+}
+
+Pipe.prototype.event = function(t, x) {
+	return this.sink.event(t, x);
+};
+
+Pipe.prototype.end = function(t, x) {
+	return this.sink.end(t, x);
+};
+
+Pipe.prototype.error = function(t, e) {
+	return this.sink.error(t, e);
+};
